requirements.txt
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.1
plotly==5.18.0
streamlit==1.28.2
langchain
langchain_experimental
langchain_openai
openai
python-dotenv
bs4==0.0.1
unstructured==0.11.2
tiktoken==0.5.2
loguru==0.7.2
watchdog==3.0.0